Make --target_branch override AutoDetectBranch
AutoDetectBranch was added to support the fact that cobalt
is normally not submitting CLs to master, which was the git
cl upload default.
Unfortunately, AutoDetectBranch does not work as expected when
working with dependent cls (git new-branch --upstream_current).
This change allows a developer to use dependent cls, and specify
--target_branch=COBALT to ensure that they are uploaded to the correct
branch.

Change-Id: I38b49c98092cf251ead1e23ace39020eb9bdfe37