Fix for branch auto-detection in GerritUpload.
The previous merge was broken.

Cobalt has extra code to auto-detect what the target branch is, since
we usually commit to the COBALT branch.

Upstream depot_tools now does detection of the remote as well, but
ends up trying to push to remote branches like refs/for/heads/remotes/COBALT

This does somehow work in Gerrit, but seems better to prefer our method.