Add a script to set the upstream git config
This script will be used as a DEPS hook to recurse
over each git repo and set the depot-tools.upstream config if
necessary (i.e. if it's different from any global setting).
This is required for git new-branch to track the correct upstream
branch.

This takes about 11s on Windows, so it caches its results in
the gclient root so that it will only re-run if the DEPS or the script
itself changes.

Change-Id: I7c1ea076458f74a448a9121c146d3caef031d9ea