Enable pyformat support on Linux
When running git cl format, use pyformat where it's available.
pyformat is installed on Goobuntu by default, and --yapf runs it
in "yapf" mode.

Hardcode it to 'full', where it reformats the entire file, because
line-mode support is not available yet. If and when upstream
depot_tools adds support for that, we can revert this.

Change-Id: I7abba1dbc75108aa498e8990d21edb39fb96f1bd