Remove unused symlinks that cause git issues.
This removes some symlinks that cause issues when a newer version of
git for windows is used to checkout the repository.

The symlinks are not used by Cobalt, but since the latest git creates
them as links while the older version of git currently included in
depot_tools treats them as files, they can cause unexpected messages
such as 'Cannot rebase: You have unstaged changes' due to 'typechange'
changes to the links that differ between the two versions of git.

Change-Id: Ie67634659fd828893e737b261e79ce485f16f81f